third_party/blink/web_tests/fast/dom/exception-in-binding-expected.txt

Tests to ensure that attributes are not set when an exceptions is raised while converting argument to JSValue.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS setName(select) threw exception Error: error.
PASS select.name is "select"
PASS setSelectedIndex(select) threw exception Error: error.
PASS select.selectedIndex is 1
PASS successfullyParsed is true

TEST COMPLETE

Value 1
Value 2
Value 3
